Add files via upload
Utilize the app.py in order to put down a stock ticker, where it displays the stock prices and uses machine learning via Natural Language Processing by looking at news articles from Yahoo News that concern the stock inputted by the user. It then gathers the sentiment scores from these websites and graphs them out, while then giving the average of the sentiment score and gives a recommendation to the user on whether to invest on the stock or not.